+++ core/src/server/state.rs
@@ -1,5 +1,6 @@
 use super::worker::RunQueueSender;
 use crate::store::SharedConnection;
+use crate::workspace_mode::WorkspaceMode;
 use std::path::PathBuf;
 
 #[derive(Clone)]
@@ -7,4 +8,5 @@ pub struct AppState {
     pub db: SharedConnection,
     pub run_queue: RunQueueSender,
     pub workspace_root: PathBuf,
+    pub workspace_mode: WorkspaceMode,
 }